requirements.txt
langchain
openai
streamlit
tiktoken
rembg
pillow
webuiapi